+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] 2022-10-23
+### Added
+- `GetTater` function to `IReader` and its implementation for generic blob reading
+- functionality to `Tater` blobs to better read object and array values
+
 ## [1.2.0] 2022-10-20
 ### Added
 - `LogError` function to `IReader` and its implementations